Various changes to fix how marc4j handles MARC records that are too large to be valid. (ie. over 99999 bytes long)
Made the MarcPermissiveStreamReader ignore the length in the leader (if need be) and read ahead to find the record terminator character.

Made the MarcStreamWriter be able to detect when a record or a field in the record would be too large to be valid, and provide an option of either writing it out anyway (with the too large lengths or offsets being capped at the maximum that will fit in the number of proscribed digits.)

Added new class CustomDecimalFormat that handles too large lengths and offsets by capping them to the maximum valid value.

Made the LeaderImpl class use the CustomDecimalFormat so that referencing portions of the leader for too large records aren't offset by 1.

Lastly added new sample data file containing an example too large record followed by two regular sized records, and modified the PermissiveReaderTest class to test the too-large-record functionality.

Obtained from: Jonathan Rochkind
Submitted by:  Robert Haschart